test: move prototype-pollution tests into select.test.ts
Fold the select() alias prototype-pollution cases into the existing
select.test.ts integration suite, reusing its createUsers() fixture,
and drop the standalone select-prototype-pollution.test.ts file.
